[+] add : my job responsive + job desktop table component
